requirements.txt
agentiq==1.1.0
aioboto3==14.3.0
aiobotocore==2.22.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aioice==0.10.1
aioitertools==0.12.0
aiortc==1.12.0
aiosignal==1.3.1
aiosqlite==0.21.0
# Editable install with no version control (aiq_semantic_kernel_demo==0.0.0)
-e /workspaces/do_aiq2/semantic_kernel_demo
aiqtoolkit==1.1.0
aiqtoolkit-langchain==1.1.0
aiqtoolkit-semantic-kernel==1.1.0rc3
alembic==1.15.2
annotated-types==0.6.0
anyio==4.9.0
appdirs==1.4.4
arize-phoenix==9.3.0
arize-phoenix-client==1.5.0
arize-phoenix-evals==0.20.6
arize-phoenix-otel==0.9.2
arxiv==2.2.0
attrs==23.2.0
Authlib==1.5.2
av==14.3.0
azure-core==1.34.0
azure-identity==1.23.0
backoff==2.2.1
beautifulsoup4==4.13.4
boto3==1.37.3
botocore==1.37.3
cachetools==5.5.2
certifi==2023.11.17
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.2.0
cloudevents==1.11.0
colorama==0.4.6
contourpy==1.3.2
cryptography==44.0.3
cycler==0.12.1
dataclasses-json==0.6.3
datasets==3.6.0
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
email_validator==2.2.0
et_xmlfile==2.0.0
expandvars==1.0.0
extratools==0.8.2.1
faiss-cpu==1.9.0
fastapi==0.115.12
feedparser==6.0.11
ffmpy==0.5.0
filelock==3.18.0
filetype==1.2.0
fonttools==4.58.0
frozenlist==1.4.1
fsspec==2025.3.0
google-crc32c==1.7.1
googleapis-common-protos==1.70.0
gradio==5.29.0
gradio_client==1.10.0
graphql-core==3.2.6
greenlet==3.0.3
# Editable install with no version control (greet==0.1.0)
-e /workspaces/do_aiq2/zagents/greet
groovy==0.1.2
grpc-interceptor==0.15.4
grpcio==1.67.1
h11==0.16.0
h2==4.2.0
hf-xet==1.1.1
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.1
hyperframe==6.1.0
idna==3.6
ifaddr==0.2.0
importlib_metadata==8.6.1
isodate==0.7.2
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.5.0
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==2.4
jsonschema==4.23.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.59
langchain-experimental==0.3.4
langchain-mcp-adapters==0.1.1
langchain-milvus==0.1.10
langchain-nvidia-ai-endpoints==0.3.10
langchain-openai==0.2.14
langchain-text-splitters==0.3.8
langgraph==0.2.76
langgraph-checkpoint==2.0.25
langgraph-sdk==0.1.69
langsmith==0.3.42
lazy-object-proxy==1.11.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.20.1
matplotlib==3.10.3
mcp==1.8.1
mdurl==0.1.2
mem0ai==0.1.98
milvus-lite==2.4.12
monotonic==1.6
more-itertools==10.7.0
msal==1.32.3
msal-extensions==1.3.1
multidict==6.0.4
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.3
openai==1.78.1
openapi-core==0.19.5
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.1
openinference-instrumentation==0.1.29
openinference-semantic-conventions==0.1.17
openpyxl==3.1.5
opentelemetry-api==1.33.0
opentelemetry-exporter-otlp==1.33.0
opentelemetry-exporter-otlp-proto-common==1.33.0
opentelemetry-exporter-otlp-proto-grpc==1.33.0
opentelemetry-exporter-otlp-proto-http==1.33.0
opentelemetry-proto==1.33.0
opentelemetry-sdk==1.33.0
opentelemetry-semantic-conventions==0.54b0
orjson==3.10.18
ormsgpack==1.9.1
packaging==23.2
pandas==2.2.3
# Editable install with no version control (pandas_expression_generator==0.1.0)
-e /workspaces/do_aiq2/zagents/pandas_expression_generator
parse==1.20.2
pathable==0.4.4
pillow==11.2.1
pkginfo==1.12.1.2
platformdirs==4.3.8
ply==3.11
portalocker==2.10.1
posthog==3.25.0
prance==23.6.21.0
prefixspan==0.5.2
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
pyarrow==20.0.0
pybars4==0.9.13
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.9.1
pydantic_core==2.27.2
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.1
PyJWT==2.10.1
pylibsrtp==0.12.0
PyMeta3==0.5.1
pymilvus==2.5.8
pyOpenSSL==25.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
# Editable install with no version control (python_express_executor==0.1.0)
-e /workspaces/do_aiq2/zagents/python_express_executor
pytz==2024.2
PyYAML==6.0.1
qdrant-client==1.14.2
ragas==0.2.15
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rpds-py==0.24.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
ruff==0.11.9
s3transfer==0.11.3
safehttpx==0.1.6
scikit-learn==1.6.1
scipy==1.15.3
semantic-kernel==1.24.1
semantic-version==2.10.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
sqlean.py==3.49.1
sse-starlette==2.3.5
starlette==0.46.2
strawberry-graphql==0.267.0
tenacity==8.2.3
threadpoolctl==3.6.0
tiktoken==0.9.0
tomlkit==0.13.2
toolz==1.0.0
tqdm==4.67.1
typer==0.15.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
urllib3==2.1.0
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.5
websockets==14.2
Werkzeug==3.1.1
wikipedia==1.4.0
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
# Editable install with no version control (z1structuredToolPandas==0.1.0)
-e /workspaces/do_aiq2/1/z1structuredToolPandas
# Editable install with no version control (z_find_plans==0.1.0)
-e /workspaces/do_aiq2/z_find_plans
# Editable install with no version control (z_mem==0.1.0)
-e /workspaces/do_aiq2/z_mem
# Editable install with no version control (z_table_agent==0.1.0)
-e /workspaces/do_aiq2/z_table_agent
# Editable install with no version control (zagent0==0.1.0)
-e /workspaces/do_aiq2/zagent0
# Editable install with no version control (zfoo==0.1.0)
-e /workspaces/do_aiq2/zfoo
zipp==3.21.0
# Editable install with no version control (zjoke==0.1.0)
-e /workspaces/do_aiq2/5/zjoke
# Editable install with no version control (zlanggraph==0.1.0)
-e /workspaces/do_aiq2/5/zlanggraph
# Editable install with no version control (zlanggraph2==0.1.0)
-e /workspaces/do_aiq2/6/zlanggraph2
zstandard==0.23.0
